Testing changes

+++ src/main/java/acme/entities/dashboard/Dashboard.java
@@ -17,7 +17,7 @@ public class Dashboard implements Serializable {
 
 	//Control check
 	
-	Integer flaggedAs;
+	Double flaggedAs;
 	Double ratioOfShouts;
 	
 	Double 	avgMoneyCurrency1;
